+++ posthog/models/session_recording_event/util.py
@@ -1,21 +1,16 @@
 import datetime
 import json
 import uuid
-from typing import Any, Dict, List, Union
+from typing import Union
 
-import pytz
 import structlog
-from dateutil.parser import isoparse
 from django.utils import timezone
 from sentry_sdk import capture_exception
 
 from posthog.client import sync_execute
 from posthog.kafka_client.client import ClickhouseProducer
 from posthog.kafka_client.topics import KAFKA_SESSION_RECORDING_EVENTS
-from posthog.models.session_recording_event.sql import (
-    BULK_INSERT_SESSION_RECORDING_EVENT_SQL,
-    INSERT_SESSION_RECORDING_EVENT_SQL,
-)
+from posthog.models.session_recording_event.sql import INSERT_SESSION_RECORDING_EVENT_SQL
 from posthog.utils import cast_timestamp_or_now
 
 logger = structlog.get_logger(__name__)
@@ -55,54 +50,6 @@ def create_session_recording_event(
         capture_exception(Exception(f"Session recording event data too large - {len(snapshot_data_json)}"))
 
     return str(uuid)
-
-
-def bulk_create_session_recording_event(events: List[Dict[str, Any]]) -> None:
-    """
-    Test only
-    """
-    # timestamp = cast_timestamp_or_now(timestamp)
-
-    inserts = []
-    params: Dict[str, Any] = {}
-    for index, event in enumerate(events):
-        timestamp = event["timestamp"]
-        if isinstance(timestamp, str):
-            timestamp = isoparse(timestamp)
-        else:
-            timestamp = timestamp.astimezone(pytz.utc)
-        timestamp = timestamp.strftime("%Y-%m-%d %H:%M:%S.%f")
-        data = {
-            "uuid": str(event["uuid"]),
-            "team_id": event["team_id"],
-            "distinct_id": event["distinct_id"],
-            "session_id": event["session_id"],
-            "window_id": event.get("window_id"),
-            "snapshot_data": json.dumps(event.get("snapshot_data", {})),
-            "timestamp": timestamp,
-            "created_at": timestamp,
-        }
-        inserts.append(
-            """(
-
-                %(uuid_{i})s,
-                %(timestamp_{i})s,
-                %(team_id_{i})s,
-                %(distinct_id_{i})s,
-                %(session_id_{i})s,
-                %(window_id_{i})s,
-                %(snapshot_data_{i})s,
-                %(created_at_{i})s,
-                now(),
-                0
-            )""".format(
-                i=index
-            )
-        )
-
-        params = {**params, **{"{}_{}".format(key, index): value for key, value in data.items()}}
-
-    sync_execute(BULK_INSERT_SESSION_RECORDING_EVENT_SQL() + ", ".join(inserts), params, flush=False)
 
 
 def get_recording_count_for_team_and_period(
